refactor(room): own ordered teardown orchestration
